separate contour + extend canvas
