+++ src/store/data_storage.ts
@@ -1,48 +1,7 @@
 import { writable, type Writable } from 'svelte/store';
 import type MapboxDraw from '@mapbox/mapbox-gl-draw';
 import { EMPTY_POLYGON_RGB } from '../lib/gl_draw_styles.js'
-
-export interface Zone {
-  type?: string,
-  id: string,
-  properties: {
-    road_lane_direction: number,
-    road_lane_num: number,
-    coordinates: [[number, number], [number, number], [number, number], [number, number]],
-    color_rgb: [number, number, number],
-    virtual_line?: {
-      geometry: [[number, number], [number, number]],
-      color_rgb: [number, number, number],
-      direction: string
-    },
-    ds_id?: string,
-    spatial_object_id?: string,
-    color_rgb_str?: string,
-  },
-  geometry: PolygonGeoJSON
-}
-
-interface PolygonGeoJSON {
-    type: string,
-    coordinates: number[][][]
-}
-
-interface ZoneFeature {
-  type: string,
-  id: string,
-  properties: {
-    road_lane_direction: number,
-    road_lane_num: number,
-    coordinates: [[number, number], [number, number], [number, number], [number, number]],
-    color_rgb: [number, number, number]
-  },
-  geometry: PolygonGeoJSON
-}
-
-export interface ZonesCollection {
-  type: string,
-  features: ZoneFeature[]
-}
+import type { Zone, ZoneFeature } from '$lib/zones.js';
 
 export const dataStorage: Writable<Map<string, Zone>> = writable(new Map<string, Zone>())
